requirements.txt
numpy 
pandas
python-dotenv # use for get/read connection info from .env
mysql-connector-python # 
pymysql # that convert data into dataFrame
scikit-learn
# -e .
